start-map/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
sqlmodel
markdownify
python-multipart
gradio==5.25.2
gradio_modal
magic_pdf==1.3.8
